Move useEvent to utils
