+++ src/useWatchSweety.ts
@@ -1,8 +1,7 @@
 import { useRef, useEffect, useCallback, useDebugValue } from "react"
 import { useSyncExternalStoreWithSelector } from "use-sync-external-store/shim/with-selector.js"
 
-import { useEvent } from "./useEvent"
-import { Compare, isEqual } from "./utils"
+import { Compare, isEqual, useEvent } from "./utils"
 import { WatchContext } from "./WatchContext"
 
 const modInc = (x: number): number => {
